Turn DuplexRequestBody into a marker interface.
And later maybe a simple public method.
